Changes from sanity branch redone
Sanity branch changes never made it. Remaking from scratch

+++ EscapeForest/Assets/Scripts/Character/SanityBar.cs
@@ -8,7 +8,7 @@ public class SanityBar : MonoBehaviour
     private BasePlayer playerScript;
     public Slider slider;
 
-    public void setSanity(int sanity)
+    public void setSanity(float sanity)
     {
         slider.value = sanity;
     }
